Added controller and updated routes
