+++ Routes/bookRoutes.js
@@ -2,29 +2,10 @@ var express= require('express');
 
 var routes=function (Book) {
     var bookRouter=express.Router();
-
+    var bookController=require('../controller/bookController.js')(Book);
     bookRouter.route('/')
-        .post(function (req,res) {
-            var book=new Book(req.body);
-            book.save();
-            res.status(201).send(book);
-
-        })
-        .get(function (req, res) {
-
-            var query={};
-
-            if(req.query.genre){
-                query.genre=req.query.genre;
-            }
-            //var responseJson={hello: "This is my API"};
-            Book.find(query, function (err, books) {
-                if(err)
-                    res.status(500).send(err);
-                else
-                    res.json(books);
-            });
-        });
+        .post(bookController.post)
+        .get(bookController.get);
     //implementing the middleware
     bookRouter.use('/:bookId', function (req, res, next) {
         Book.findById(req.params.bookId, function (err, book) {
